Create ArrayCheck
We first take the values from the user and then check if the size of array was same, as we need to find if two arrays are identical, it should have same size. Then we take the element of A and check whether if it can be found in B, if it found then we swap it with the last element and we decrease the size of B, so we don't search same element again and again.